requirements.txt
streamlit
google-generativeai
python-dotenv
# pytesseract
# pillow
# transformers
# librosa